Cleanup

+++ src/main/java/b_model/entities/Settings.java
@@ -1,19 +1,19 @@
 package b_model.entities;
 
 public class Settings {
-	private final float targetTemperature;
+	private final double targetTemperature;
 	private final int temperatureMargin;
 	private final int humidityThreshold;
 	private final int co2Threshold;
 
-	public Settings(float targetTemperature, int temperatureMargin, int humidityThreshold, int co2Threshold) {
+	public Settings(double targetTemperature, int temperatureMargin, int humidityThreshold, int co2Threshold) {
 		this.targetTemperature = targetTemperature;
 		this.temperatureMargin = temperatureMargin;
 		this.humidityThreshold = humidityThreshold;
 		this.co2Threshold = co2Threshold;
 	}
 
-	public float getTargetTemperature() {
+	public double getTargetTemperature() {
 		return targetTemperature;
 	}
